@@ -10,13 +10,15 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  google_maps_flutter: ^2.5.0
+  google_maps_flutter: ^2.6.0 # or latest version
   google_fonts: ^6.1.0
+  cached_network_image: ^3.3.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^3.0.1
+  geolocator: ^12.0.0
 
 flutter:
   uses-material-design: true
